+++ doc/source/plot-header.py
@@ -1,11 +1,10 @@
 import matplotlib.pyplot as plt
-import numpy as np
+from pns.util import circular_data, unit_sphere
 
-from skpns import PNS
-from skpns.util import circle, circular_data, unit_sphere
+from skpns import ExtrinsicPNS
 
-pns = PNS(n_components=2)
-X = circular_data()
+pns = ExtrinsicPNS(n_components=2)
+X = circular_data([0, -1, 0])
 X_new = pns.fit_transform(X)
 v, r = pns.v_[0], pns.r_[0]
 
@@ -13,8 +12,4 @@ fig = plt.figure()
 ax = fig.add_subplot(projection="3d", computed_zorder=False)
 ax.plot_surface(*unit_sphere(), color="skyblue", alpha=0.6, edgecolor="gray")
 ax.scatter(*X.T, marker=".", color="tab:blue")
-ax.plot(*circle(v, r), color="tab:orange")
-ax.scatter(*pns.to_hypersphere(X_new).T, marker="x", color="tab:green")
-
-xs, ys, zs = X.T
-ax.set_box_aspect((np.ptp(xs), np.ptp(ys), np.ptp(zs)))
+ax.scatter(*pns.inverse_transform(X_new).T, marker="x", color="tab:green")