+++ src/skpns/__init__.py
@@ -7,18 +7,23 @@ Low-level functions are available in :mod:`skpns.pns`.
        "Analysis of principal nested spheres." Biometrika 99.3 (2012): 551-568.
 """
 
-from .sklearn import PNS, ExtrinsicPNS, IntrinsicPNS
+from .sklearn import ExtrinsicPNS, IntrinsicPNS, InverseExtrinsicPNS
 
 __all__ = [
     "ExtrinsicPNS",
-    "PNS",
     "IntrinsicPNS",
+    "InverseExtrinsicPNS",
 ]
 
 try:
     from skl2onnx import update_registered_converter
 
-    from .onnx import extrinsicpns_converter, intrinsicpns_converter, shape_calculator
+    from .onnx import (
+        extrinsicpns_converter,
+        intrinsicpns_converter,
+        inverse_extrinsicpns_converter,
+        shape_calculator,
+    )
 
     update_registered_converter(
         IntrinsicPNS,
@@ -31,6 +36,12 @@ try:
         "SkpnsExtrinsicPNS",
         shape_calculator,
         extrinsicpns_converter,
+    )
+    update_registered_converter(
+        InverseExtrinsicPNS,
+        "SkpnsInverseExtrinsicPNS",
+        shape_calculator,
+        inverse_extrinsicpns_converter,
     )
 
 except ModuleNotFoundError: