+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - UNRELEASED
+
+Transformers are now based on [pnspy](https://pypi.org/project/pnspy/) package.
+
+### Added
+
+- `InverseExtrinsicPNS` class is added.
+
+### Removed
+
+- `PNS` alias for `ExtrinsicPNS` is removed.
+- `ExtrinsicPNS.to_hypersphere()` is removed. Use `inverse_transform()` method instead.
+
+- `skpns.pns` module is removed. Use `PnsPy` package (`pns.pns`, `pns.pss`, `pns.base`) instead.
+- `skpns.util` module is removed. Use `PnsPy` package (`pns.util`) instead.
+
 ## [1.3.0] - 2025-11-26
 
 ### Added